% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussfit.R
\name{gaussfunc}
\alias{gaussfunc}
\title{Gaussian Function}
\usage{
gaussfunc(x, mu, sigma)
}
\arguments{
\item{x}{the vector of values at which to evaluate the Gaussian}

\item{mu}{the center of the Gaussian}

\item{sigma}{the spread of the Gaussian (must be greater than 0)}
}
\value{
vector of values of the Gaussian
}
\description{
This function returns the unnormalized (height of 1.0) Gaussian curve with a
given center and spread.
}
\examples{
x = seq(-4, 4, length.out = 100)
y = gaussfunc(x, 0, 1)
plot(x, y)

}
