% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgrv.R
\name{HG1}
\alias{HG1}
\title{Evaluate the First-Degree Generalized Hermite-Gaussian Function}
\usage{
HG1(x, mu, sig)
}
\arguments{
\item{x}{the vector of values at which to evaluate the function}

\item{mu}{the center parameter of the function}

\item{sig}{the spread parameter of the function}
}
\value{
vector of values of the specified first-degree generalized Hermite-Gaussian function
}
\description{
This function evaluates the first-degree Hermite-Gaussian function with a
general center and spread.
}
\examples{
x = seq(50, 60, length.out=100)
y = HG1(x, 55, 1)
plot(x, y)

}
