% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{observed_spec}
\alias{observed_spec}
\title{Observed spectrum for the star 51 Pegasi (HD 217014)}
\format{
A dataframe with 628 rows and the following 3 columns:
\describe{
\item{Wavelength}{the wavelength of the spectrum, in Angstroms}
\item{Flux}{normalized flux of the spectrum, unitless}
\item{Uncertainty}{the uncertainty of the flux measurements, unitless}
...
}
}
\source{
\url{https://arxiv.org/abs/2003.08851}
}
\usage{
observed_spec
}
\description{
A small portion of one observed spectrum collected by EXPRES
\strong{Petersburg et. al (2020)}.
}
\keyword{datasets}
