% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spectra}
\alias{spectra}
\title{Observed spectra for the star 51 Pegasi (HD 217014)}
\format{
A list with 56 elements, each of which has 2 variables:
\describe{
  \item{Wavelength}{the wavelength of the spectrum, in Angstroms}
  \item{Flux}{normalized flux of the spectrum, unitless}
  ...
}
}
\source{
\url{https://arxiv.org/abs/2003.08851}
}
\usage{
spectra
}
\description{
56 observed spectra as collected by EXPRES \strong{Petersburg et. al (2020)}.
Only the subset of the spectrum between 5000 and 5005 Angstroms is given here.
}
\keyword{datasets}
