% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template}
\alias{template}
\title{Estimated template spectrum for the star 51 Pegasi (HD 217014)}
\format{
A data frame with 1893 rows and 2 variables:
\describe{
  \item{Wavelength}{the wavelength of the spectrum, in Angstroms}
  \item{Flux}{normalized flux of the spectrum, unitless}
  ...
}
}
\source{
\url{https://arxiv.org/abs/2003.08851}
}
\usage{
template
}
\description{
A small portion of the estimated template produced with the method of
\href{https://arxiv.org/abs/2005.14083}{Holzer et. al (2020)} on a set 
of 55 observed spectra from EXPRES.
}
\keyword{datasets}
